Port to 1.21.6

+++ src/main/java/dev/gigaherz/toolbelt/client/radial/TextRadialMenuItem.java
@@ -37,7 +37,7 @@ public class TextRadialMenuItem extends RadialMenuItem
         String textString = text.getString();
         float x = context.x - context.font.width(textString) / 2.0f;
         float y = context.y - context.font.lineHeight / 2.0f;
-        context.graphics.drawString(context.font, textString, x, y, color, true);
+        context.graphics.drawString(context.font, textString, (int)x, (int)y, color, true);
     }
 
     @Override
